+++ apps/erp/package.json
@@ -16,7 +16,6 @@
     "@carbon/documents": "workspace:*",
     "@carbon/ee": "workspace:*",
     "@carbon/env": "workspace:*",
-    "@carbon/field-help": "workspace:*",
     "@carbon/form": "workspace:*",
     "@carbon/glossary": "workspace:*",
     "@carbon/jobs": "workspace:*",
