--- packages/field-help/tsconfig.json
@@ -1,6 +1,0 @@
-{
-  "$schema": "https://json.schemastore.org/tsconfig",
-  "extends": "@carbon/config/tsconfig/react-library.json",
-  "include": ["src"],
-  "exclude": ["**/*.d.ts", "./dist", "node_modules"]
-}
