+++ packages/form/src/components/Radios.tsx
@@ -1,9 +1,9 @@
-import { LabelWithHelp } from "@carbon/field-help";
 import type { TermId } from "@carbon/glossary";
 import {
   FormControl,
   FormErrorMessage,
   FormLabel,
+  LabelWithHelp,
   RadioGroup,
   RadioGroupItem
 } from "@carbon/react";
