+++ packages/form/src/components/TextareaControlled.tsx
@@ -1,4 +1,3 @@
-import { LabelWithHelp } from "@carbon/field-help";
 import type { TermId } from "@carbon/glossary";
 import type { TextareaProps } from "@carbon/react";
 import {
@@ -6,6 +5,7 @@ import {
   FormErrorMessage,
   FormHelperText,
   FormLabel,
+  LabelWithHelp,
   Textarea as TextAreaBase
 } from "@carbon/react";
 import type { ChangeEvent } from "react";
